+++ Orion+#/Source/Server/Source/Modules/S_Types.cs
@@ -37,7 +37,7 @@
             public byte Sex;
             public byte Classes;
             public int Sprite;
-            public byte Level;
+            public int Level;
             public int Exp;
 
             public byte Pk;
@@ -46,9 +46,9 @@
             public int[] Vital;
 
             // Stats
-            public byte[] Stat;
+            public int[] Stat;
 
-            public byte Points;
+            public int Points;
 
             // Worn equipment
             public int[] Equipment;
